Support a permissions boundary on the create-role operation
This charm operates, in part, by creating roles conferring desired
permissions. AWS IAM provides a "permissions boundary" mechanism by
which a user or role granted permissions may be bounded.

https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_boundaries.html

Allow configuration of a permissions boundary policy so that a user or
role may be granted permission to create a role, but the potential
powers of such a role may be bounded.

For example, add a condition such as the following to the principal's
iam:CreateRole permission:

    "Condition": {
      "StringEquals": {
        "iam:PermissionsBoundary":
        "arn:aws:iam::...account...:policy/JujuIntegratorBoundaries"
      }
    }

and create the JujuIntegratorBoundaries policy to include the merger
of policies to be used from this charm's files/policies/*.json; see
charms.layer.aws.enable_* for which policy file each enable_* request
uses. Configure an application deployed from this charm with
permissions-boundary-policy=JujuIntegratorBoundaries (or with that
policy's full ARN).